Working Nonogram Generator/Puzzle Interface
It's not the prettiest thing in the world but it works. This small
front-page app lets you click on blocks to make your own nonogram,
outputted in object format, or load up a set of nonograms I programmed
in. I'd like to get this so that it load puzzles from a mongo database
and let you login with Facebook, then save/edit puzzles you've already
made, but this is a good start. This will probably end up being a very
specialized node module.